Added calculations for the one sample t test

+++ src/lexer.h
@@ -88,7 +88,7 @@ int lex_end_of_command (void);
 int lex_integer_p (void);
 long lex_integer (void);
 int lex_double_p (void);
-long lex_double (void);
+double lex_double (void);
 
 /* Token matching functions. */
 int lex_match (int);
